move Link back to public api
