+++ geode-management/src/main/java/org/apache/geode/management/api/ConfigurationResult.java
@@ -22,7 +22,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import org.apache.geode.annotations.Experimental;
 import org.apache.geode.management.configuration.AbstractConfiguration;
-import org.apache.geode.management.internal.Links;
+import org.apache.geode.management.configuration.Links;
 import org.apache.geode.management.runtime.RuntimeInfo;
 
 /**
