+++ geode-management/src/main/java/org/apache/geode/management/configuration/AbstractConfiguration.java
@@ -23,7 +23,6 @@ import org.apache.geode.annotations.Experimental;
 import org.apache.geode.lang.Identifiable;
 import org.apache.geode.management.api.ClusterManagementService;
 import org.apache.geode.management.api.JsonSerializable;
-import org.apache.geode.management.internal.Links;
 import org.apache.geode.management.runtime.RuntimeInfo;
 
 /**
