+++ geode-management/src/main/java/org/apache/geode/management/configuration/GatewayReceiver.java
@@ -19,7 +19,6 @@ import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import org.apache.geode.management.internal.Links;
 import org.apache.geode.management.runtime.GatewayReceiverInfo;
