+++ geode-management/src/main/java/org/apache/geode/management/configuration/Region.java
@@ -23,7 +23,6 @@ import java.util.stream.Collectors;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import org.apache.geode.management.internal.Links;
 import org.apache.geode.management.runtime.RuntimeRegionInfo;
 
 /**
