+++ geode-management/src/main/java/org/apache/geode/management/internal/ClientClusterManagementService.java
@@ -15,8 +15,8 @@
 
 package org.apache.geode.management.internal;
 
+import static org.apache.geode.management.configuration.Links.URI_VERSION;
 import static org.apache.geode.management.internal.Constants.INCLUDE_CLASS_HEADER;
-import static org.apache.geode.management.internal.Links.URI_VERSION;
 
 import java.util.Date;
 import java.util.concurrent.CompletableFuture;
