+++ geode-web-management/src/integrationTest/java/org/apache/geode/management/internal/rest/RebalanceIntegrationTest.java
@@ -15,7 +15,7 @@
 
 package org.apache.geode.management.internal.rest;
 
-import static org.apache.geode.management.internal.Links.URI_VERSION;
+import static org.apache.geode.management.configuration.Links.URI_VERSION;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.core.StringContains.containsString;
